@@ -1,13 +1,13 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Bogus" Version="27.0.1" />
-    <PackageReference Include="Grpc.Core" Version="1.21.0" />
+    <PackageReference Include="Bogus" Version="28.4.4" />
+    <PackageReference Include="Grpc.Core" Version="2.27.0" />
   </ItemGroup>
 
   <ItemGroup>
