@@ -5,13 +5,14 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using ProgramaPontos.Infra.Ioc.AspNetCore;
 using Swashbuckle.AspNetCore.Swagger;
+using Swashbuckle.Swagger;
 
 namespace ProgramaPontos.API
 {
     public class Startup
     {
 
-        private string version = "1.0";
+        private readonly string version = "1.0";
         private readonly string applicationName = "ProgramaPontos";
 
         public Startup(IConfiguration configuration)
@@ -24,24 +25,22 @@ namespace ProgramaPontos.API
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddMvc(s=>s.EnableEndpointRouting=false);
 
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc(version, new Info { Title = "ProgramaPontos", Version = version });
+                c.SwaggerDoc(version, new Microsoft.OpenApi.Models.OpenApiInfo { Title = "ProgramaPontos",Version= version });
             });
+
+            
 
             services.AddProgramaPontosServices(Configuration);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
+           
             app.UseMvc();
 
             app.UseSwagger();
